EES-4830 Move `EnumUtil` to Utils directory
