+++ src/GovUk.Education.ExploreEducationStatistics.Common/Converters/EnumToEnumValueJsonConverter.cs
@@ -1,6 +1,6 @@
 using System;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
-using GovUk.Education.ExploreEducationStatistics.Common.Services;
+using GovUk.Education.ExploreEducationStatistics.Common.Utils;
 using Newtonsoft.Json;
 
 namespace GovUk.Education.ExploreEducationStatistics.Common.Converters
