+++ src/GovUk.Education.ExploreEducationStatistics.Common/Utils/EnumUtil.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
 
-namespace GovUk.Education.ExploreEducationStatistics.Common.Services
+namespace GovUk.Education.ExploreEducationStatistics.Common.Utils
 {
     public static class EnumUtil
     {
