+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model.Tests/Extensions/DataImportStatusExtensionTests.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using GovUk.Education.ExploreEducationStatistics.Common.Services;
+using GovUk.Education.ExploreEducationStatistics.Common.Utils;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Content.Model.DataImportStatus;
